implementation/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
distro==1.9.0
Django==5.0.6
dotenv==0.0.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
image==1.5.33
openai==1.33.0
paho-mqtt==2.1.0
pillow==10.3.0
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
RPi.GPIO==0.7.1
six==1.16.0
smbus==1.1.post2
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
